Whitepaper of OAH Family
OAH Family, a group of enthusiasts who imagined to connect the Old Age Homes with the compact world of blockchain cryptocurrency, announced today that they will be airdropping 1,100,000 OAH Coins, to join the initiative by volunteering the platform on social media. OAH Coin started with the vision to raise funds for Old Age Homes, which is a cryptocurrency based on blockchain technology supporting PoW (Proof of Work) and PoS (Proof of Stake) algorithms for a Green Cryptocurrency.

OAH Family intends to use OAH Coin as the primary connection to provide the requisite facilities to the elderly and in the process create a growth avenue for our volunteering investors. It can be said that a simple bridge between the aged and young population, with a vision of “society for all ages”, is designed to provide volunteering opportunities for youth to create a new architecture for aging, with promising returns for both the aged and young generation.

To participate in the airdrop program one needs to become a volunteer by:

• Signing up on the official website- https://oah.family/volunteer-signup or by following the invitation link
• Joining telegram group and the channel will reward with 0.5$ and 0.5$ worth OAH coin respectively
• On Twitter, by following, liking, or retweeting the pinned post, an OAH coin worth 1$ will be rewarded
• On Facebook - Like @oahfamilyofficial pages, Like Share and comment on the latest post would be enough to get another OAH coin worth 1$
• On Instagram - Follow @oah.family on Instagram, like comment and sharing latest post, will reward you OAH coin of another dollar
• On Medium - Follow OAHFAMILY would reward OAH coin of 1 $
• On Reddit - Upvoting to add karma on Reddit OAHFAMILY post would reward you the same
• Similarly it goes with LinkedIn and Discord
• Subscribing YouTube channel of OAHFAMILY and liking and sharing would reward one OAH coin worth 2$.

"To provide 20x growth to volunteering investors by mid-2022, a 100x growth by EOY 2023 and 1000x growth by EOY 2025." This statement by the group seems very promising and worth investing in.